Fixed path error in bower.json
https://github.com/bower/bower.json-spec

The "main" file should be a path to the file, not just a filename. I am using your plugin along with my build tools and it's failing due to this path not being correct.
I have filled similar PRs for lots of other bower plugins before. This seems like a common mistake.

I have also updated to version number, so all you need to do is merge and tag a new release. Which you can do from the github website :) so I'm hoping you can get this in ASAP as in my source code I want to remove the hack I've had to put in place.

Many thanks