Allow calling of jQBrowser if jQuery is not defined
This required a tad bit of restructuring, mainly making the browser
matching a part of the function and not outside of it. This also
prevents the code from running twice if jQuery is loaded and the
function is called.

It may be that we want the function to be called on page load
regardless. In that case, the changes required are trivial.